UPDATE: Appointment and DoctorList
Appointment management is now it's own dashboard and the doctor list is now specific to the actual clinics available doctors. Still fixing how the queue and appointment pull's it's information! -R-- is done, still implementing the C-UD of CRUD
